Move RouteTarget back to server.startup
