+++ flow-server/src/main/java/com/vaadin/flow/router/NavigationStateBuilder.java
@@ -18,7 +18,7 @@ package com.vaadin.flow.router;
 import java.util.List;
 
 import com.vaadin.flow.component.Component;
-import com.vaadin.flow.router.internal.RouteTarget;
+import com.vaadin.flow.server.startup.RouteTarget;
 
 /**
  * A builder class for constructing new {@link NavigationState} instances.
