+++ flow-server/src/main/java/com/vaadin/flow/router/internal/NavigationRouteTarget.java
@@ -19,6 +19,7 @@ import java.io.Serializable;
 import java.util.Map;
 
 import com.vaadin.flow.router.UrlParameters;
+import com.vaadin.flow.server.startup.RouteTarget;
 
 /**
  * Contains the information resulted from searching a route target using a
