+++ flow-server/src/main/java/com/vaadin/flow/server/RouteRegistry.java
@@ -32,7 +32,7 @@ import com.vaadin.flow.router.RouterLayout;
 import com.vaadin.flow.router.RoutesChangedListener;
 import com.vaadin.flow.router.UrlParameters;
 import com.vaadin.flow.router.internal.NavigationRouteTarget;
-import com.vaadin.flow.router.internal.RouteTarget;
+import com.vaadin.flow.server.startup.RouteTarget;
 import com.vaadin.flow.shared.Registration;
 
 /**
@@ -127,7 +127,7 @@ public interface RouteRegistry extends Serializable {
      */
     RouteTarget getRouteTarget(Class<? extends Component> target,
             UrlParameters parameters);
-    
+
     /**
      * Gets the optional navigation target class for a given path. Returns an
      * empty optional if no navigation target corresponds to the given url.
